refactor: 타입 정의 표준화 및 도메인별 분리 완료
- 각 도메인별로 타입 정의 분리 (auth, practice, team)
- 공통 타입은 shared/types로 이동 (Role, Gender, MemberStatus 등)
- BaseEntity, TimestampedEntity 유틸리티 타입 추가
- 도메인별 barrel export에 타입 포함
- 타입 일관성 및 재사용성 향상
- Predictability 가이드라인 적용
