+++ frontend/src/domains/auth/index.ts
@@ -2,3 +2,4 @@
 export * from './components';
 export * from './hooks/AuthContext';
 export * from './api/authApi';
+export * from './types';
