+++ frontend/src/domains/practice/index.ts
@@ -1,3 +1,4 @@
 // Practice domain barrel exports
 export * from './components';
 export * from './api/practiceApi';
+export * from './types';
