+++ frontend/src/domains/team/index.ts
@@ -1,2 +1,3 @@
 // Team domain barrel exports
 export * from './api/teamApi';
+export * from './types';
